+++ packages/front/src/routes/depot-scdl/components/Step4/Step4.svelte
@@ -40,22 +40,29 @@
 
         <div class="fr-col-12 fr-col-md-4">
             <InfoBox title="Besoin d’aide ?">
-                <p>
-                    Vous pouvez vous reporter à notre documentation pour plus d’informations.
-                    <br />
+                <ul>
+                    <li class="fr-mb-2v">
+                        Consulter la documentation officielle pour plus d’informations :
+                        <br />
+                        <TargetBlankLink href="https://schema.data.gouv.fr/scdl/subventions/">
+                            Voir la documentation
+                        </TargetBlankLink>
+                    </li>
+                    <!--                    <li class="fr-mb-2v">-->
+                    <!--                        Préparer votre fichier en vous appuyant sur-->
+                    <!--                        <TargetBlankLink href="https://schema.data.gouv.fr/scdl/subventions/">-->
+                    <!--                            notre modèle SCDL &lt;!&ndash; todo : en attente lien vers le gabarit xlsx + chgt blank link par download link &ndash;&gt;-->
+                    <!--                        </TargetBlankLink>-->
+                    <!--                    </li>-->
+                    <li class="fr-mb-2v">Nous contacter via la bulle de chat</li>
+                    <li>
+                        Participer à notre webinaire pour plus d’accompagnement dans la mise à jour de votre fichier :
+                    </li>
                     <TargetBlankLink
-                        href="https://www.notion.so/R-gles-de-format-SCDL-1281788663a380e1a57efdd9b324c1ba">
-                        Voir la documentation
+                        href="https://datasubvention.beta.gouv.fr/permanence-scdl-creer-structurer-et-deposer-vos-donnees/">
+                        S'inscrire
                     </TargetBlankLink>
-                    <br />
-                </p>
-                <p class="fr-mb-4v">
-                    Vous pouvez également nous contacter via la bulle de chat ou participer à nos permanences mensuelles
-                    avec notre équipe. Nous vous guiderons dans la mise à jour de votre fichier.
-                    <br />
-                    <TargetBlankLink href="mailto:{ctrl.contactEmail}">S'inscrire à la permanence</TargetBlankLink>
-                    <!-- todo : quel lien ?-->
-                </p>
+                </ul>
             </InfoBox>
         </div>
     </div>